Avoid request processor unnecessarily accessing app data
While it's nice to scope use of app data, it's desirable to not have
request processor pulling out data that may not even be used by the
downstream dependency.

In this case, we can avoid having to populate some fake app data in the
test that is passed through to our mock OJP interface and never used.